Extract DFSTools to propagation.go.  Export necessary fields and methods.
