+++ internal/pkg/source/source.go
@@ -44,14 +44,12 @@ type classifier interface {
 // Source.sanitized notes sanitizer calls that sanitize this Source
 type Source struct {
 	Node         ssa.Node
-	marked       map[ssa.Node]bool
-	preOrder     []ssa.Node
-	sanitizers   []*sanitizer.Sanitizer
-	config       classifier
-	taggedFields fieldtags.ResultType
+	Marked       map[ssa.Node]bool
+	PreOrder     []ssa.Node
+	Sanitizers   []*sanitizer.Sanitizer
+	Config       classifier
+	TaggedFields fieldtags.ResultType
 }
-
-type DFSTools Source
 
 // Pos returns the token position of the SSA Node associated with the Source.
 func (s *Source) Pos() token.Pos {
@@ -75,221 +73,12 @@ func (s *Source) Pos() token.Pos {
 
 // New constructs a Source
 func New(in ssa.Node, config classifier, taggedFields fieldtags.ResultType) *Source {
-	dfsTool := &DFSTools{
+	return &Source{
 		Node:         in,
-		marked:       make(map[ssa.Node]bool),
-		config:       config,
-		taggedFields: taggedFields,
+		Marked:       make(map[ssa.Node]bool),
+		Config:       config,
+		TaggedFields: taggedFields,
 	}
-
-	src := Source(*dfsTool)
-	return &src
-}
-
-// dfs performs Depth-First-Search on the def-use graph of the input Source.
-// While traversing the graph we also look for potential sanitizers of this Source.
-// If the Source passes through a sanitizer, dfs does not continue through that Node.
-func (s *DFSTools) Dfs(n ssa.Node, maxInstrReached map[*ssa.BasicBlock]int, lastBlockVisited *ssa.BasicBlock, isReferrer bool) {
-	if s.shouldNotVisit(n, maxInstrReached, lastBlockVisited, isReferrer) {
-		return
-	}
-	s.preOrder = append(s.preOrder, n)
-	s.marked[n] = true
-
-	mirCopy := map[*ssa.BasicBlock]int{}
-	for m, i := range maxInstrReached {
-		mirCopy[m] = i
-	}
-
-	if instr, ok := n.(ssa.Instruction); ok {
-		instrIndex, ok := indexInBlock(instr)
-		if !ok {
-			return
-		}
-
-		if mirCopy[instr.Block()] < instrIndex {
-			mirCopy[instr.Block()] = instrIndex
-		}
-
-		lastBlockVisited = instr.Block()
-	}
-
-	s.visit(n, mirCopy, lastBlockVisited)
-}
-
-func (s *DFSTools) shouldNotVisit(n ssa.Node, maxInstrReached map[*ssa.BasicBlock]int, lastBlockVisited *ssa.BasicBlock, isReferrer bool) bool {
-	if s.marked[n] {
-		return true
-	}
-
-	if !hasTaintableType(n) {
-		return true
-	}
-
-	if instr, ok := n.(ssa.Instruction); ok {
-		instrIndex, ok := indexInBlock(instr)
-		if !ok {
-			return true
-		}
-
-		// If the referrer is in a different block from the one we last visited,
-		// and it can't be reached from the block we are visiting, then stop visiting.
-		if lastBlockVisited != nil && instr.Block() != lastBlockVisited && !s.canReach(lastBlockVisited, instr.Block()) {
-			return true
-		}
-
-		// If this call's index is lower than the highest seen so far in its block,
-		// then this call is "in the past". If this call is a referrer,
-		// then we would be propagating taint backwards in time, so stop traversing.
-		// (If the call is an operand, then it is being used as a value, so it does
-		// not matter when the call occurred.)
-		if _, ok := instr.(*ssa.Call); ok && instrIndex < maxInstrReached[instr.Block()] && isReferrer {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (s *DFSTools) visit(n ssa.Node, maxInstrReached map[*ssa.BasicBlock]int, lastBlockVisited *ssa.BasicBlock) {
-	if s.Node == n {
-		s.visitReferrers(n, maxInstrReached, lastBlockVisited)
-		return
-	}
-
-	switch t := n.(type) {
-	case *ssa.Alloc:
-		// An Alloc represents the allocation of space for a variable. If a Node is an Alloc,
-		// and the thing being allocated is not an array, then either:
-		// a) it is a Source value, in which case it will get its own traversal when sourcesFromBlocks
-		//    finds this Alloc
-		// b) it is not a Source value, in which case we should not visit it.
-		// However, if the Alloc is an array, then that means the source that we are visiting from
-		// is being placed into an array, slice or varags, so we do need to keep visiting.
-		if _, isArray := utils.Dereference(t.Type()).(*types.Array); isArray {
-			s.visitReferrers(n, maxInstrReached, lastBlockVisited)
-		}
-
-	case *ssa.Call:
-		if callee := t.Call.StaticCallee(); callee != nil && s.config.IsSanitizer(utils.DecomposeFunction(callee)) {
-			s.sanitizers = append(s.sanitizers, &sanitizer.Sanitizer{Call: t})
-		}
-
-		// This is to avoid attaching calls where the source is the receiver, ex:
-		// core.Sinkf("Source id: %v", wrapper.Source.GetID())
-		if recv := t.Call.Signature().Recv(); recv != nil && isSourceType(s.config, s.taggedFields, recv.Type()) {
-			return
-		}
-
-		s.visitReferrers(n, maxInstrReached, lastBlockVisited)
-		for _, a := range t.Call.Args {
-			if canBeTaintedByCall(a.Type()) {
-				s.Dfs(a.(ssa.Node), maxInstrReached, lastBlockVisited, false)
-			}
-		}
-
-	case *ssa.FieldAddr:
-		deref := utils.Dereference(t.X.Type())
-		typPath, typName := utils.DecomposeType(deref)
-		fieldName := utils.FieldName(t)
-		if !s.config.IsSourceField(typPath, typName, fieldName) && !s.taggedFields.IsSourceFieldAddr(t) {
-			return
-		}
-		s.visitReferrers(n, maxInstrReached, lastBlockVisited)
-		s.visitOperands(n, maxInstrReached, lastBlockVisited)
-
-	// Everything but the actual integer Index should be visited.
-	case *ssa.Index:
-		s.visitReferrers(n, maxInstrReached, lastBlockVisited)
-		s.Dfs(t.X.(ssa.Node), maxInstrReached, lastBlockVisited, false)
-
-	// Everything but the actual integer Index should be visited.
-	case *ssa.IndexAddr:
-		s.visitReferrers(n, maxInstrReached, lastBlockVisited)
-		s.Dfs(t.X.(ssa.Node), maxInstrReached, lastBlockVisited, false)
-
-	// Only the Addr (the Value that is being written to) should be visited.
-	case *ssa.Store:
-		s.Dfs(t.Addr.(ssa.Node), maxInstrReached, lastBlockVisited, false)
-
-	// Only the Map itself can be tainted by an Update.
-	// The Key can't be tainted.
-	// The Value can propagate taint to the Map, but not receive it.
-	// MapUpdate has no referrers, it is only an Instruction, not a Value.
-	case *ssa.MapUpdate:
-		s.Dfs(t.Map.(ssa.Node), maxInstrReached, lastBlockVisited, false)
-
-	// The only Operand that can be tainted by a Send is the Chan.
-	// The Value can propagate taint to the Chan, but not receive it.
-	// Send has no referrers, it is only an Instruction, not a Value.
-	case *ssa.Send:
-		s.Dfs(t.Chan.(ssa.Node), maxInstrReached, lastBlockVisited, false)
-
-	// These nodes' operands should not be visited, because they can only receive
-	// taint from their operands, not propagate taint to them.
-	case *ssa.BinOp, *ssa.ChangeInterface, *ssa.ChangeType, *ssa.Convert, *ssa.Extract, *ssa.MakeChan, *ssa.MakeMap, *ssa.MakeSlice, *ssa.Phi, *ssa.Range:
-		s.visitReferrers(n, maxInstrReached, lastBlockVisited)
-
-	// These nodes don't have operands; they are Values, not Instructions.
-	case *ssa.Const, *ssa.FreeVar, *ssa.Global, *ssa.Lookup, *ssa.Parameter:
-		s.visitReferrers(n, maxInstrReached, lastBlockVisited)
-
-	// These nodes don't have referrers; they are Instructions, not Values.
-	case *ssa.Go:
-		s.visitOperands(n, maxInstrReached, lastBlockVisited)
-
-	// These nodes are both Instructions and Values, and currently have no special restrictions.
-	case *ssa.Field, *ssa.MakeInterface, *ssa.Select, *ssa.Slice, *ssa.TypeAssert, *ssa.UnOp:
-		s.visitReferrers(n, maxInstrReached, lastBlockVisited)
-		s.visitOperands(n, maxInstrReached, lastBlockVisited)
-
-	// These nodes cannot propagate taint.
-	case *ssa.Builtin, *ssa.DebugRef, *ssa.Defer, *ssa.Function, *ssa.If, *ssa.Jump, *ssa.MakeClosure, *ssa.Next, *ssa.Panic, *ssa.Return, *ssa.RunDefers:
-
-	default:
-		fmt.Printf("unexpected node received: %T %v; please report this issue\n", n, n)
-	}
-}
-
-func (s *DFSTools) visitReferrers(n ssa.Node, maxInstrReached map[*ssa.BasicBlock]int, lastBlockVisited *ssa.BasicBlock) {
-	if n.Referrers() == nil {
-		return
-	}
-	for _, r := range *n.Referrers() {
-		s.Dfs(r.(ssa.Node), maxInstrReached, lastBlockVisited, true)
-	}
-}
-
-func (s *DFSTools) visitOperands(n ssa.Node, maxInstrReached map[*ssa.BasicBlock]int, lastBlockVisited *ssa.BasicBlock) {
-	for _, o := range n.Operands(nil) {
-		if *o == nil {
-			continue
-		}
-		s.Dfs((*o).(ssa.Node), maxInstrReached, lastBlockVisited, false)
-	}
-}
-
-func (s *DFSTools) canReach(start *ssa.BasicBlock, dest *ssa.BasicBlock) bool {
-	if start.Dominates(dest) {
-		return true
-	}
-
-	stack := stack([]*ssa.BasicBlock{start})
-	seen := map[*ssa.BasicBlock]bool{start: true}
-	for len(stack) > 0 {
-		current := stack.pop()
-		if current == dest {
-			return true
-		}
-		for _, s := range current.Succs {
-			if seen[s] {
-				continue
-			}
-			seen[s] = true
-			stack.push(s)
-		}
-	}
-	return false
 }
 
 // compress removes the elements from the graph that are not required by the
@@ -298,7 +87,7 @@ func (s *DFSTools) canReach(start *ssa.BasicBlock, dest *ssa.BasicBlock) bool {
 // is, it is also removed.
 func (s *Source) compress() []ssa.Node {
 	var compressed []ssa.Node
-	for _, n := range s.preOrder {
+	for _, n := range s.PreOrder {
 		switch n.(type) {
 		case *ssa.Call:
 			compressed = append(compressed, n)
@@ -310,12 +99,12 @@ func (s *Source) compress() []ssa.Node {
 
 // HasPathTo returns true when a Node is part of declaration-use graph.
 func (s *Source) HasPathTo(n ssa.Node) bool {
-	return s.marked[n]
+	return s.Marked[n]
 }
 
 // IsSanitizedAt returns true when the Source is sanitized by the supplied instruction.
 func (s *Source) IsSanitizedAt(call ssa.Instruction) bool {
-	for _, san := range s.sanitizers {
+	for _, san := range s.Sanitizers {
 		if san.Dominates(call) {
 			return true
 		}
@@ -359,7 +148,7 @@ func identify(conf classifier, ssaInput *buildssa.SSA, taggedFields fieldtags.Re
 func sourcesFromParams(fn *ssa.Function, conf classifier, taggedFields fieldtags.ResultType) []*Source {
 	var sources []*Source
 	for _, p := range fn.Params {
-		if isSourceType(conf, taggedFields, p.Type()) {
+		if IsSourceType(conf, taggedFields, p.Type()) {
 			sources = append(sources, New(p, conf, taggedFields))
 		}
 	}
@@ -371,7 +160,7 @@ func sourcesFromClosure(fn *ssa.Function, conf classifier, taggedFields fieldtag
 	for _, p := range fn.FreeVars {
 		switch t := p.Type().(type) {
 		case *types.Pointer:
-			if isSourceType(conf, taggedFields, t) {
+			if IsSourceType(conf, taggedFields, t) {
 				sources = append(sources, New(p, conf, taggedFields))
 			}
 		}
@@ -408,7 +197,7 @@ func sourcesFromBlocks(fn *ssa.Function, conf classifier, taggedFields fieldtags
 			// have an Alloc and we'll miss it.
 			case *ssa.Extract:
 				t := v.Tuple.Type().(*types.Tuple).At(v.Index).Type()
-				if _, ok := t.(*types.Pointer); ok && isSourceType(conf, taggedFields, t) {
+				if _, ok := t.(*types.Pointer); ok && IsSourceType(conf, taggedFields, t) {
 					sources = append(sources, New(v, conf, taggedFields))
 				}
 				continue
@@ -428,7 +217,7 @@ func sourcesFromBlocks(fn *ssa.Function, conf classifier, taggedFields fieldtags
 			}
 
 			// all of the instructions that the switch lets through are values as per ssa/doc.go
-			if v := instr.(ssa.Value); isSourceType(conf, taggedFields, v.Type()) {
+			if v := instr.(ssa.Value); IsSourceType(conf, taggedFields, v.Type()) {
 				sources = append(sources, New(v.(ssa.Node), conf, taggedFields))
 			}
 		}
@@ -436,7 +225,7 @@ func sourcesFromBlocks(fn *ssa.Function, conf classifier, taggedFields fieldtags
 	return sources
 }
 
-func hasTaintableType(n ssa.Node) bool {
+func HasTaintableType(n ssa.Node) bool {
 	if v, ok := n.(ssa.Value); ok {
 		switch t := v.Type().(type) {
 		case *types.Basic:
@@ -448,20 +237,20 @@ func hasTaintableType(n ssa.Node) bool {
 	return true
 }
 
-func isSourceType(c classifier, tf fieldtags.ResultType, t types.Type) bool {
+func IsSourceType(c classifier, tf fieldtags.ResultType, t types.Type) bool {
 	deref := utils.Dereference(t)
 	switch tt := deref.(type) {
 	case *types.Named:
-		return c.IsSourceType(utils.DecomposeType(tt)) || isSourceType(c, tf, tt.Underlying())
+		return c.IsSourceType(utils.DecomposeType(tt)) || IsSourceType(c, tf, tt.Underlying())
 	case *types.Array:
-		return isSourceType(c, tf, tt.Elem())
+		return IsSourceType(c, tf, tt.Elem())
 	case *types.Slice:
-		return isSourceType(c, tf, tt.Elem())
+		return IsSourceType(c, tf, tt.Elem())
 	case *types.Chan:
-		return isSourceType(c, tf, tt.Elem())
+		return IsSourceType(c, tf, tt.Elem())
 	case *types.Map:
-		key := isSourceType(c, tf, tt.Key())
-		elem := isSourceType(c, tf, tt.Elem())
+		key := IsSourceType(c, tf, tt.Key())
+		elem := IsSourceType(c, tf, tt.Elem())
 		return key || elem
 	case *types.Struct:
 		return hasTaggedField(tf, tt)
@@ -481,20 +270,20 @@ func isSourceType(c classifier, tf fieldtags.ResultType, t types.Type) bool {
 // A type can be tainted by a call if it is itself a pointer or pointer-like type (according to
 // pointer.CanPoint), or it is an array/struct that holds an element that can be tainted by
 // a call.
-func canBeTaintedByCall(t types.Type) bool {
+func CanBeTaintedByCall(t types.Type) bool {
 	if pointer.CanPoint(t) {
 		return true
 	}
 
 	switch tt := t.(type) {
 	case *types.Array:
-		return canBeTaintedByCall(tt.Elem())
+		return CanBeTaintedByCall(tt.Elem())
 
 	case *types.Struct:
 		for i := 0; i < tt.NumFields(); i++ {
 			// this cannot cause an infinite loop, because a struct
 			// type cannot refer to itself except through a pointer
-			if canBeTaintedByCall(tt.Field(i).Type()) {
+			if CanBeTaintedByCall(tt.Field(i).Type()) {
 				return true
 			}
 		}
@@ -532,7 +321,7 @@ func isProducedBySanitizer(v ssa.Value, conf classifier) bool {
 }
 
 // indexInBlock returns this instruction's index in its parent block.
-func indexInBlock(target ssa.Instruction) (int, bool) {
+func IndexInBlock(target ssa.Instruction) (int, bool) {
 	for i, instr := range target.Block().Instrs {
 		if instr == target {
 			return i, true
@@ -543,9 +332,9 @@ func indexInBlock(target ssa.Instruction) (int, bool) {
 	return 0, false
 }
 
-type stack []*ssa.BasicBlock
+type Stack []*ssa.BasicBlock
 
-func (s *stack) pop() *ssa.BasicBlock {
+func (s *Stack) Pop() *ssa.BasicBlock {
 	if len(*s) == 0 {
 		log.Println("tried to pop from empty stack")
 	}
@@ -554,6 +343,6 @@ func (s *stack) pop() *ssa.BasicBlock {
 	return popped
 }
 
-func (s *stack) push(b *ssa.BasicBlock) {
+func (s *Stack) Push(b *ssa.BasicBlock) {
 	*s = append(*s, b)
 }
